feat: refactor dashboard structure by creating CompanyDashboard component and implementing suspense loading
